Crate cirrus_client_executor

Subspace Executor

Executors, a separate class of nodes in addition to the consensus nodes (farmers) in Subspace, are designed to reduce the burden of maintaining the chain state for farmers by decoupling the consensus and computation. As an execution layer, executor chain itself does no rely on any typical blockchain consensus like PoW for producing blocks, the block production of executor chain is totally driven by the consensus layer which are collectively maintained by Subspace farmers. Please refer to the white paper Computation section for more in-depth description and analysis.

Specifically, executors are responsible for producing a Bundle on each slot from the primary chain and producing an ExecutionReceipt on each primary block.

On each new primary chain slot, executors will collect a set of extrinsics from the transaction pool which are verified to be able to cover the transaction fee, and then use these extrinsics to create a Bundle, submitting it to the primary chain. The submitted bundles are mere blob from the point of primary chain.

On each imported primary block, executors will extract all the bundles from the primary block and convert the bundles to a list of extrinsics, construct a custom BlockBuilder to build a secondary block. The execution trace of all the extrinsics and hooks like initialize_block/finalize_block will be recorded during the block execution. Once the secondary block has been imported successfully, an executor that wins the election for producing an execution receipt will publish the receipt over the executors network.

The execution receipt of each block contains all the intermediate state roots during the block execution, which will be gossiped in the executor network. All executors whether running as an authority or a full node will compute each block and generate an execution receipt independently, once the execution receipt received from the network does not match the one produced locally, a FraudProof will be generated and reported to the primary chain accordingly.

Notes

Currently, the following terms are interexchangeable in the executor context:

Farmer, consensus node.

Executor, execution/compute node.

Primary chain, consensus layer.

Secondary chain, execution layer.

Structs

Executor

The implementation of the Cirrus Executor.

Enums

GossipMessageError

Error type for cirrus gossip handling.